+++ src/main/java/ru/yandex/practicum/filmorate/dao/LikeDao.java
@@ -9,31 +9,31 @@ public interface LikeDao {
      * к фильму с id == filmId
      * от пользователя с id == userId
      */
-    void saveLike(int filmId, int userId);
+    void saveLike(Long filmId, Long userId);
 
     /**
      * Удалить лайк из БД
      * к фильму с id == filmId
      * от пользователя с id == userId
      */
-    void deleteLike(int filmId, int userId);
+    void deleteLike(Long filmId, Long userId);
 
     /**
      * Получить количество лайков
      * к фильму по его id
      */
-    Integer getCountLikesByFilmId(int filmId);
+    Integer getCountLikesByFilmId(Long filmId);
 
     /**
      * Получить список id пользователей
      * поставивших лайк к фильму по его id
      */
-    List<Integer> getAllUsersLikeFilm(int film_id);
+    List<Integer> getAllUsersLikeFilm(Long film_id);
 
     /**
      * Проверка наличия лайка в БД
      * у фильма с id == filmId
      * от пользователя с id == userId
      */
-    boolean containsInStorage(int filmId, int userId);
+    boolean containsInStorage(Long filmId, Long userId);
 }
